% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEC_features.R
\name{asymptotic_eigp}
\alias{asymptotic_eigp}
\title{Asymptotic Wald's test for testing the exponent in a EIGP model.}
\usage{
asymptotic_eigp(data, eta0 = 1, theta1, eta1)
}
\arguments{
\item{data}{n by 1 vector with all positive entries.}

\item{eta0}{To test if the exponent equals 1, the default for eta0 is et to be 1.}

\item{theta1}{The unrestricted MLE of theta.}

\item{eta1}{The unrestricted MLE of eta.}
}
\value{
This function returns the test statistic and the p-value of the Wald's test.
}
\description{
This function computes the test statistic and the p-value of Wald's test for the exponent parameter in EIGP model.
}
\details{
asymptotic_eigp
}
\examples{
sample1 = eigp_sampling(1000,eta = 1.1,theta = 3)
theta1 = mle_search_eigp(data = sample1)$theta
eta1 = mle_search_eigp(data = sample1)$eta
asymptotic_eigp(sample1,eta0 = 1,theta1,eta1)
}
