% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEC_features.R
\name{cdf_eep}
\alias{cdf_eep}
\title{The cumulative distribution function of EEP.}
\usage{
cdf_eep(theta, eta, data)
}
\arguments{
\item{theta}{The location parameter for the base distribution (eta = 1). The value needs to be positive.}

\item{eta}{The exponent parameter. The value provided needs to be positive.}

\item{data}{The data item.}
}
\value{
Return the cumulative probability of EEP at the specific location.
}
\description{
cdf_eep
}
\examples{
cdf_eep(1,2,5)
}
