% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEC_estimation.R
\name{eep_nll}
\alias{eep_nll}
\title{The EEP Negative Log-likelihood Function.}
\usage{
eep_nll(x, m, data)
}
\arguments{
\item{x}{Vector of parameters.}

\item{m}{The number of data items less than the density change point.}

\item{data}{Observations.}
}
\value{
A scalar that represents the negative loglikelihood of a EEP sample given the model parameter and the data.
}
\description{
This function serves as the objective function for the Maximum Likelihood Estimation procedure for EEP.
}
\details{
eep_nll
}
\examples{
eep_nll(c(2,2),50,seq(1:100))
}
