% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEC_estimation.R
\name{eep_optim}
\alias{eep_optim}
\title{The Wrapper Function that Returns the Final Estimates from Maximum Likelihood Estimation for EEP.}
\usage{
eep_optim(data, init = c(1, 1), lower_bound = c(0.01, 0.01))
}
\arguments{
\item{data}{Observations.}

\item{init}{The vector of initial values of the model parameters. The default is c(1,1).}

\item{lower_bound}{The vector of the lower bound for the parameters. The default is c(0.01,0.01).}
}
\value{
A data frame with 1 row and 3 columns that contains the MLE of theta, eta, and the predicted density change point.
}
\description{
This function serves as a wrapper that returns the final estimates of theta, eta, and the corresponding density change point
}
\details{
eep_optim
}
\examples{
eep_optim(seq(1:100))
}
