% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEC_sampling.R
\name{eep_sampling}
\alias{eep_sampling}
\title{The Random Number Generation Function for EIGP}
\usage{
eep_sampling(n, theta, eta)
}
\arguments{
\item{n}{Number of observations. (n>=1)}

\item{theta}{The location parameter for the parent EP distribution (eta = 1). The value needs to be positive.}

\item{eta}{The exponent parameter. The parameter should be positive.}
}
\value{
returns a numerical vector of size n.
}
\description{
Create a EEP random sample.
}
\details{
eep_sampling
}
\examples{
eep_sampling(100,1,1)
}
