% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEC_estimation.R
\name{eigp_nll}
\alias{eigp_nll}
\title{The EIGP Negative Log-likelihood Function.}
\usage{
eigp_nll(x, m, data)
}
\arguments{
\item{x}{Vector of parameters.}

\item{m}{The number of data items less than the density change point.}

\item{data}{Observations.}
}
\value{
A scalar that represents the negative loglikelihood of a EIGP sample given the model parameter and the data.
}
\description{
This function serves as the objective function for the Maximum Likelihood Estimation procedure for EIGP.
}
\details{
eigp_nll
}
\examples{
eigp_nll(c(2,2),50,seq(1:100))
}
