% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEC_sampling.R
\name{eigp_sampling}
\alias{eigp_sampling}
\title{The Random Number Generation Function for EIGP}
\usage{
eigp_sampling(n, theta, eta)
}
\arguments{
\item{n}{Number of observations. (n>=1)}

\item{theta}{The location parameter for the parent IGP distribution (eta = 1). The value needs to be positive.}

\item{eta}{The exponent parameter. The parameter should be positive.}
}
\value{
This function returns a numerical vector of size n.
}
\description{
Create a EIGP random sample.
}
\details{
eigp_sampling
}
\examples{
eigp_sampling(100,1,1)
}
