% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEC_features.R
\name{exp_eep}
\alias{exp_eep}
\title{The negative log density of a sample item if it follows exponential in a EEP model}
\usage{
exp_eep(x, theta, eta)
}
\arguments{
\item{x}{The value of a sample item.}

\item{theta}{The location parameter for the base distribution (eta = 1). The value needs to be positive.}

\item{eta}{The exponent parameter. The value provided needs to be positive.}
}
\value{
This function return the negative log density of a sample item if if it follows exponential in a EEP model.
}
\description{
This function return the negative log density of a sample item if if it follows exponential in a EEP model.
}
\details{
exp_exp
}
\examples{
exp_eep(1,5,2)
}
