% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEC_features.R
\name{hazard_eep}
\alias{hazard_eep}
\title{The hazard function of EEP.}
\usage{
hazard_eep(theta, eta, data)
}
\arguments{
\item{theta}{The location parameter for the base distribution (eta = 1). The value needs to be positive.}

\item{eta}{The exponent parameter. The value provided needs to be positive.}

\item{data}{The data item.}
}
\value{
Return the hazard of EEP at the specific location.
}
\description{
hazard_eep
}
\examples{
hazard_eep(2,1,5)
plot(hazard_eep(2,1,seq(0.01,100,by=0.01)))
}
