% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEC_features.R
\name{mle_eep}
\alias{mle_eep}
\title{Analytical solution of theta given eta in EEP model.}
\usage{
mle_eep(s, m, n)
}
\arguments{
\item{s}{a numeric value the sum of log(1/x_i^eta), where i is from 1 to m.}

\item{m}{m is the number of data items less than the density change point.}

\item{n}{n is the sample size, n has to be greater than m.}
}
\value{
This function returns the Maximum Likelihood Estimate of theta for a given eta
}
\description{
This function provides the analytical solution of theta for given eta EEP model.
}
\details{
mle_eep
}
\examples{
mle_eep(5,2,5)
}
