% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEC_features.R
\name{mle_iter_eigp}
\alias{mle_iter_eigp}
\title{Iteration function to find the analytical solution of theta given eta and data in EIGP model.}
\usage{
mle_iter_eigp(data, eta)
}
\arguments{
\item{data}{n by 1 vector with all positive entries.}

\item{eta}{The exponent parameter. This value is greater than 0.}
}
\value{
This function returns the Maximum Likelihood Estimate of theta for a given eta with data.
}
\description{
This function finds the analytical solution of theta given eta and data in EIGP model.
}
\details{
mle_iter_eigp
}
\examples{
mle_iter_eigp(seq(1:100),2)
}
