% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEC_features.R
\name{neg_log_eep}
\alias{neg_log_eep}
\title{The negative log likelihood function for EEP distribution.}
\usage{
neg_log_eep(y, theta, eta)
}
\arguments{
\item{y}{n by 1 vector with all positive entries.}

\item{theta}{The location parameter for the base distribution (eta = 1). The value needs to be positive.}

\item{eta}{The exponent parameter. The value provided needs to be positive.}
}
\value{
This function return the negative log density of a sample item if if it follows Pareto in a EEP model.
}
\description{
This function computes the negative log-likelihood for EEP distribution.
}
\details{
neg_log_eigp
}
\examples{
neg_log_eep(seq(1:100),2,2)
}
