% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEC_features.R
\name{pareto_eigp}
\alias{pareto_eigp}
\title{The negative log density of a sample item if it follows Pareto in a EIGP model}
\usage{
pareto_eigp(x, theta, eta)
}
\arguments{
\item{x}{The value of a sample item.}

\item{theta}{The location parameter for the base distribution (eta = 1). The value needs to be positive.}

\item{eta}{The exponent parameter. The value provided needs to be positive.}
}
\value{
This function return the negative log density of a sample item if if it follows Pareto in a EIGP model.
}
\description{
This function return the negative log density of a sample item if if it follows Pareto in a EIGP model.
}
\details{
pareto_eigp
}
\examples{
pareto_eigp(10,5,2)
}
