% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEC_features.R
\name{pdf_eep}
\alias{pdf_eep}
\title{The probability function of EEP.}
\usage{
pdf_eep(theta, eta, data)
}
\arguments{
\item{theta}{The location parameter for the base distribution (eta = 1). The value needs to be positive.}

\item{eta}{The exponent parameter. The value provided needs to be positive.}

\item{data}{The data item.}
}
\value{
Return the density of EEP
}
\description{
pdf_eep
}
\examples{
pdf_eep(1,2,5)
}
