% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEC_features.R
\name{pdf_eigp}
\alias{pdf_eigp}
\title{The probability density function of EIGP.}
\usage{
pdf_eigp(theta, eta, data)
}
\arguments{
\item{theta}{The location parameter for the base distribution (eta = 1). The value needs to be positive.}

\item{eta}{The exponent parameter. The value provided needs to be positive.}

\item{data}{The data item.}
}
\value{
Return the density of EIGP
}
\description{
pdf_eigp
}
\examples{
pdf_eigp(1,2,5)
}
