% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEC_features.R
\name{q_eigp}
\alias{q_eigp}
\title{The quantile function of EIGP.}
\usage{
q_eigp(theta, eta, p)
}
\arguments{
\item{theta}{The location parameter for the base distribution (eta = 1). The value needs to be positive.}

\item{eta}{The exponent parameter. The value provided needs to be positive.}

\item{p}{This indicates the p-th percentile. p is greater than 0 and less than 100.}
}
\value{
Return the p-th percentile of EIGP.
}
\description{
q_eigp
}
\examples{
q_eigp(1,2,5)
}
