% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEC_features.R
\name{se_eep}
\alias{se_eep}
\title{The function for calculating the standard errors of the parameters of EEP model.}
\usage{
se_eep(data, theta, eta)
}
\arguments{
\item{data}{Observations.}

\item{theta}{The MLE of theta}

\item{eta}{The MLE of eta}
}
\value{
The estimate of SE for theta and eta
}
\description{
This function find the parameter estimates of EEP through a grid search procedure.
}
\details{
se_eep
}
\examples{
sample1 = eep_sampling(200,eta = 2,theta = 3)
theta = mle_search_eep(data = sample1)$theta
eta = mle_search_eep(data = sample1)$eta
se_eep(sample1,theta,eta)
}
