% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEC_features.R
\name{se_eigp}
\alias{se_eigp}
\title{The function for calculating the standard errors of the parameters of EIGP model.}
\usage{
se_eigp(data, theta, eta)
}
\arguments{
\item{data}{Observations.}

\item{theta}{The MLE of theta}

\item{eta}{The MLE of eta}
}
\value{
The estimate of SE for theta and eta
}
\description{
This function find the parameter estimates of EIGP through a grid search procedure.
}
\details{
se_eigp
}
\examples{
sample1 = eigp_sampling(200,eta = 2,theta = 3)
theta = mle_search_eigp(data = sample1)$theta
eta = mle_search_eigp(data = sample1)$eta
se_eigp(sample1,theta,eta)
}
