% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEC_estimation.R
\name{validation}
\alias{validation}
\title{The validation Function for Model Parameters.}
\usage{
validation(data, estimate)
}
\arguments{
\item{data}{Observations.}

\item{estimate}{The data frame with 2 columns named 'theta' and 'eta'.}
}
\value{
A Boolean vector.
}
\description{
This function checks if the estimates from raw_est_eigp or raw_est_eep satisfy the pre-defined conditions for the parameters.
}
\details{
validation
}
\examples{
estimate = raw_est_eigp(seq(1:100),init = c(1,1),lower_bound = c(0.01,0.01))
estimate = data.frame(estimate)
colnames(estimate) = c('theta','eta')
validation(seq(1:100),estimate)
}
