% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEC_estimation.R
\name{eep_nll}
\alias{eep_nll}
\title{The EEP negative log-likelihood function.}
\usage{
eep_nll(x, m, data)
}
\arguments{
\item{x}{A 2 by 1 vector.}

\item{m}{m is the number of data items less than the density change point}

\item{data}{n by 1 vector with all positive entries}
}
\description{
This function serves as the objective function for the Maximum Likelihood Estimation procedure for EEP.
}
\details{
eep_nll

x is a 2 by 1 vector; m denotes the number of data items less than the density change point; data is a n by 1 vector, where n denotes the sample size of the data.
}
\examples{
eep_nll(c(2,2),50,seq(1:100))
}
