% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEC_estimation.R
\name{eep_optim}
\alias{eep_optim}
\title{The wrapper function that returns the final estimates from Maximum Likelihood Estimation.}
\usage{
eep_optim(data, init = c(1, 1), lower_bound = c(0.01, 0.01))
}
\arguments{
\item{data}{a n by 1 vector with all positive entries.}

\item{init}{a 2 by 1 vector serves as the initial values of the model parameters. The default is c(1,1).}

\item{lower_bound}{a 2 by 1 vector serves as the lower bound of the parameters. The default is c(0.01,0.01).}
}
\value{
a data frame with 1 row and 3 columns that contains the MLE of theta, eta, and the predicted density change point.
}
\description{
This function serves as a wrapper that returns the final estimates of theta, eta, and the corresponding density change point
}
\details{
eigp_optim
}
\examples{
eep_optim(seq(1:100))
}
