% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEC_sampling.R
\name{eep_sampling}
\alias{eep_sampling}
\title{Sampling from EEP distribution.}
\usage{
eep_sampling(n, theta, eta)
}
\arguments{
\item{n}{A positive integer to specify the sample size}

\item{theta}{The location parameter for the base distribution (eta = 1). The value needs to be positive.}

\item{eta}{The exponent parameter. The value provided needs to be positive.}
}
\value{
returns a numerical vector of size n.
}
\description{
Create a EEP random sample of size n, with parameters theta and eta.
}
\details{
eep_sampling

Input an the sample size as n, parameters theta and eta, returns a numerical vector of size n.
}
\examples{
eep_sampling(100,1,1)
}
