% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEC_estimation.R
\name{raw_est_eigp}
\alias{raw_est_eigp}
\title{The optimization function for EIGP maximum likelihood estimation.}
\usage{
raw_est_eigp(data, init = c(1, 1), lower_bound = c(0.01, 0.01))
}
\arguments{
\item{data}{a n by 1 vector with all positive entries.}

\item{init}{a 2 by 1 vector serves as the initial values of the model parameters. The default is c(1,1).}

\item{lower_bound}{a 2 by 1 vector serves as the lower bound of the parameters. The default is c(0.01,0.01).}
}
\value{
a n-1 by 2 matrix with estimates of theta and eta for n-1 different locations of density change points (1st column for theta, 2nd column for eta).
}
\description{
This function serves as the optimization function for EIGP at different locations of density change points.
}
\details{
raw_est_eigp

x is a 2 by 1 vector; m denotes the number of data items less than the density change point; data is a n by 1 vector, where n denotes the sample size of the data.
}
\examples{
raw_est_eigp(seq(1:100))
}
