% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEC_estimation.R
\name{validation}
\alias{validation}
\title{The validation function for model parameters.}
\usage{
validation(data, estimate)
}
\arguments{
\item{data}{a n by 1 vector with all positive entries.}

\item{estimate}{a data frame with 2 columns named 'theta' and 'eta'.}
}
\value{
a n-1 by 1 Boolean vector.
}
\description{
This function checks if the estimates from raw_est_eigp or raw_est_eep satisfy the pre-defined conditions for the parameters.
}
\details{
validation
}
\examples{
estimate = raw_est_eigp(seq(1:100),init = c(1,1),lower_bound = c(0.01,0.01))
estimate = data.frame(estimate)
colnames(estimate) = c('theta','eta')
validation(seq(1:100),estimate)
}
